% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_pq.R
\name{is_gappa_installed}
\alias{is_gappa_installed}
\title{Is gappa available?}
\usage{
is_gappa_installed(path = NULL)
}
\arguments{
\item{path}{Optional path to the gappa executable. Default to NULL, in which
case it is looked up in three places, in order: the \code{phylopq.gappapath}
option, a copy installed by \code{\link[=install_gappa]{install_gappa()}}, then the system \code{PATH}.}
}
\value{
A logical of length one.
}
\description{
\if{html}{\out{<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">}} \if{html}{\out{<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental">}}\if{html}{\out{</a>}}

Check whether the gappa placement-analysis program used by
\code{\link[=assign_placement_pq]{assign_placement_pq()}} is installed. Useful to guard examples, tests and
vignette chunks.
}
\examples{
is_gappa_installed()
}
\seealso{
\code{\link[=assign_placement_pq]{assign_placement_pq()}}, \code{\link[=install_gappa]{install_gappa()}}, \code{\link[=is_epang_installed]{is_epang_installed()}}
}
\author{
Adrien Taudière
}
