% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopq-package.R
\docType{package}
\name{phylopq-package}
\alias{phylopq}
\alias{phylopq-package}
\title{phylopq: Phylogenetic Analysis for 'phyloseq' Objects}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A phylogenetic analysis layer for 'phyloseq' objects within the 'pqverse' ecosystem. Provides helpers for tree construction from taxonomy tables, phylogenetic distance metrics, phylogeny-aware ordinations and visualisations, and integration with external phylogenetic placement tools. Built on top of 'phyloseq', 'ape' and 'MiscMetabar'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adrientaudiere.github.io/phylopq/}
  \item \url{https://github.com/adrientaudiere/phylopq}
  \item Report bugs at \url{https://github.com/adrientaudiere/phylopq/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
