% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_package_pq.R
\name{print.phylopq_refpkg}
\alias{print.phylopq_refpkg}
\title{Print a reference package}
\usage{
\method{print}{phylopq_refpkg}(x, ...)
}
\arguments{
\item{x}{A \code{phylopq_refpkg} object, as returned by \code{\link[=ref_package_pq]{ref_package_pq()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a reference package
}
\examples{
\donttest{
library(MiscMetabar)
data(data_fungi_mini)
df <- subset_taxa_pq(data_fungi_mini, taxa_sums(data_fungi_mini) > 9000)
ref_taxa <- phyloseq::taxa_names(df)[1:8]
refpkg <- ref_package_pq(
  phyloseq::prune_taxa(ref_taxa, df),
  tree = ape::rtree(length(ref_taxa), tip.label = ref_taxa)
)
print(refpkg)
}
}
\seealso{
\code{\link[=ref_package_pq]{ref_package_pq()}}
}
\author{
Adrien Taudière
}
